Move Label to its own include.
