+++ vm/assembler.h
@@ -36,6 +36,7 @@
 #include <stddef.h>
 #include <stdint.h>
 #include <limits.h>
+#include "label.h"
 
 class AssemblerBase
 {
@@ -183,164 +184,6 @@ class ExternalAddress : public AddressValue
   explicit ExternalAddress(void* p)
    : AddressValue(p)
   {}
-};
-
-// A label is a lightweight object to assist in managing relative jumps. It
-// exists in three states:
-//   * Unbound, Unused: The label has no incoming jumps, and its position has
-//     not yet been fixed in the instruction stream.
-//   * Unbound, Used: The label has not yet been fixed at a position in the
-//     instruction stream, but it has incoming jumps.
-//   * Bound: The label has been fixed at a position in the instruction stream.
-//
-// When a label is unbound and used, the offset stored in the Label is a linked
-// list threaded through each individual jump. When the label is bound, each
-// jump instruction in this list is immediately patched with the correctly
-// computed relative distance to the label.
-//
-// We keep sizeof(Label) == 4 to make it embeddable within code streams if
-// need be (for example, SourcePawn mirrors the source code to maintain jump
-// maps).
-class Label
-{
-  // If set on status_, the label is bound.
-  static const int32_t kBound = (1 << 0);
-
- public:
-  Label()
-   : status_(0)
-  {
-  }
-  ~Label()
-  {
-    assert(!used() || bound());
-  }
-
-  static inline bool More(uint32_t status) {
-    return status != 0;
-  }
-  static inline uint32_t ToOffset(uint32_t status) {
-    return status >> 1;
-  }
-
-  bool used() const {
-    return bound() || !!(status_ >> 1);
-  }
-  bool bound() const {
-    return !!(status_ & kBound);
-  }
-  uint32_t offset() const {
-    assert(bound());
-    return ToOffset(status_);
-  }
-  uint32_t status() const {
-    assert(!bound());
-    return status_;
-  }
-  uint32_t addPending(uint32_t pc) {
-    assert(pc <= INT_MAX / 2);
-    uint32_t prev = status_;
-    status_ = pc << 1;
-    return prev;
-  }
-  void bind(uint32_t offset) {
-    assert(!bound());
-    status_ = (offset << 1) | kBound;
-    assert(this->offset() == offset);
-  }
-
- protected:
-  // Note that 0 as an invalid offset is okay, because the offset we save for
-  // pending jumps are after the jump opcode itself, and therefore 0 is never
-  // valid, since there are no 0-byte jumps.
-  uint32_t status_;
-};
-
-// Label that suppresses its assert, for non-stack use.
-class SilentLabel : public Label
-{
- public:
-  SilentLabel()
-  {}
-  ~SilentLabel() {
-    status_ = 0;
-  }
-};
-
-// A CodeLabel is a special form of Label intended for absolute that
-// are within the code buffer, and thus aren't known yet, and will be
-// automatically fixed up later.
-// 
-// Unlike normal Labels, these do not store a list of incoming uses.
-class CodeLabelBase
-{
-  // If set on status_, the label is bound.
-  static const int32_t kBound = (1 << 0);
-
- public:
-  CodeLabelBase()
-   : status_(0)
-  {
-  }
-  ~CodeLabelBase()
-  {
-    assert(!used() || bound());
-  }
-
-  static inline uint32_t ToOffset(uint32_t status) {
-    return status >> 1;
-  }
-
-  bool used() const {
-    return bound() || !!(status_ >> 1);
-  }
-  bool bound() const {
-    return !!(status_ & kBound);
-  }
-  uint32_t offset() const {
-    assert(bound());
-    return ToOffset(status_);
-  }
-  uint32_t status() const {
-    assert(!bound());
-    return status_;
-  }
-  void use(uint32_t pc) {
-    assert(!used());
-    status_ = (pc << 1);
-    assert(ToOffset(status_) == pc);
-  }
-  void bind(uint32_t offset) {
-    assert(!bound());
-    status_ = (offset << 1) | kBound;
-    assert(this->offset() == offset);
-  }
-
- protected:
-  uint32_t status_;
-};
-
-// Absolute address, any pointer size. These are fixed up when calling
-// emitToExecutableMemory().
-class CodeLabel : public CodeLabelBase
-{
-};
-
-class SilentCodeLabel : public CodeLabel
-{
- public:
-  ~SilentCodeLabel() {
-    status_ = 0;
-  }
-};
-
-class PatchLabel : public SilentCodeLabel
-{
-};
-
-// A 32-bit offset from the start of the code segment.
-class OffsetLabel : public CodeLabelBase
-{
 };
 
 #endif // _include_sourcepawn_assembler_h__
